requirements.txt
fastapi[standard]
uvicorn
sqlmodel
psycopg2-binary  # For PostgreSQL connection (required for Supabase)
python-dotenv
pytest
pydantic
pydantic-settings
pytest-html
pytest-asyncio
pytest-cov
python-jose[cryptography]
python-dotenv==1.0.0
httpx